api/requirements.txt
langchain
langchain-classic
langchain-openai
langchain-core
langchain_community
langchain-huggingface
langchain-text-splitters
docx2txt
pypdf
langchain_chroma
python-multipart
streamlit>=1.33
streamlit-chat
streamlit-extras
chinese-calendar
lunardate
sentence-transformers
fastapi
uvicorn